Speed improvements for `integration_tests.yml` and CLI docker targets changes
Fixes #6735

`cli/Dockerfile` has been refactored to have different possible final targets, one per os/arch, while keeping the old `multi-arch` target that builds for everything.

The `DOCKER_MULTIARCH` env var has been replaced with `DOCKER_TARGET`, that should match any of the targets in that Dockerfile. If not set, its value is set to the host's os/arch automatically in `bin/_docker.sh`.

`bin/_docker.sh` is consumed by `bin/docker-build-cli-bin`, whose logic is now simpler and allows to be called through something like `DOCKER_TARGET=xxx bin/docker-build-cli-bin` to be able to build the CLI inside docker for a specific os/arch (again, if `DOCKER_TARGET` is unset, it'll build for the host's os/arch).

`bin/docker-pull-binaries` was also simplified to always pull all the binaries regardless of `DOCKER_TARGET`.

The `integration_tests.yml` workflow now sets `DOCKER_TARGET=linux-amd64` because that's all that's required. This makes the `Docker build (cli-bin)` job to no longer be the lengthiest one, which results in a speedup of 5mins for the `docker_build` part of the workflow.

The `release.yml` continues to work as before, having now `DOCKER_TARGET=multi-arch`, given that besides the `linux-amd64` CLI we also need `linux-arm64` and the windows CLI to be available there.