policy-controller: Make kubelet IP indexing optional
In some cluster environments--namely Amazon EKS, nodes do not include a
`podCIDRs` configuration, so it's impossible to automatically detect a
node's Kubelet IP.

This change modifies node indexing to support this case by handling the
case when neither `podCIDR` nor `podCIDRs` is set. In this case, no
`default:kubelet` authorization is set on authorization responses. This
means that unauthenticated kubelet traffic will need to be explicitly
enabled in these clusters when restrictive default policies are used.

Fixes #6742