Add sp-validator ports to inbound port configuration (#6814)
The destination controller sets a list of its inbound ports for policy
discovery, but this list currently omits the sp-validator's ports. When
using a default-deny policy, this prevents the sp-validator from
starting properly.

This change updates the destination controller's default list of ports
to include 8443 and 9997 for the sp-validator as well as 9443 for the
policy controller's admission controller.