Create a top-level Cargo workspace
Various development tools (including Rust Analyzer and some reusable
actions) expect the root of the project to define a Cargo workspace.

In order to work more naturally with these tools, this change moves the
`Cargo.lock`, `rust-toolchain`, and `deny.toml` files to the root of the
project. A `Cargo.toml` is factored out of `policy-controller` to define
the top-level workspace.