actions: Do not run tests on push to main
Each time we push to the main branch, we run a full set of unit and
integration tests, even though the branch has already passed CI before
merging. When we have many branches in-flight, this work chews into our
actions usage, delaying actions for open PRs that are more important to
run expediently.

This change disables running these tests on push to `main`. We continue
to run `CodeQL` checks on main, as these checks appear to be necessary.