opaque ports check: consider service and pod ports when checking annotation values (#6801)
After #6719 merged the proxy injector started adding a default list for the `config.linkerd.io/opaque-ports` annotation to pods and services if they and their namespace do not have the annotation already.

Addresses part of #6783

For the control plane, the default list that pods get can be different than the services that select them. For example, the `linkerd-sp-validator` exposes `443`—which is opaque by default—but it targets `8443` on `linkerd-destination` pods which is not opaque by default. Therefore, `linkerd-sp-validator`'s default list is `443` and `linkerd-destination`'s default list is `8443`.

The way that the opaque ports check currently works is it naively issues warnings when a service and a pod that it selects have different values for the annotation. With this change, it now accounts for the service's `port` and `targetPort` and the pod's `containerPort`.

---

The main change about the opaque ports check is it now considers `service.Spec.Ports` and `pod.Spec.Containers.Ports`. Using those port lists, it can now figure out when a service targets—by integer or name—an opaque port on a pod and warn the user if it is not annotated on the pod.

Additionally, it can look at a pods's opaque port list and warn the user if a service that targets any of the ports—again by integer or name—does not have that port marked as opaque.

This solves the problem above about a service marking `443` as opaque, but a pod that it selects marking `8443` as opaque. The check can now see that `443 == sp-validator`, and `sp-validator == 8443`  on the pod.

I've updated the unit tests and added some additional ones that exercise the named port behavior.

Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>