proxy: v2.159.0
This release includes a change to how inbound connections are forwarded
on the local pod: instead of sending traffic to the application on
127.0.0.1, the proxy now only forwards traffic on the pod's public IP
address. This protects services that are only bound on the loopback
interface from being exposed to other pods.

This release also improves authorization metrics to include TLS labels
(including the client's identity, if available).

The `/tasks` admin endpoint has been removed, as it no longer works. It
will be replaced with an admin endpoint supporting tokio's `console` in
a future release.

---

* build(deps): bump tracing-subscriber from 0.2.20 to 0.2.22 (linkerd/linkerd2-proxy#1261)
* build(deps): bump tracing from 0.1.26 to 0.1.27 (linkerd/linkerd2-proxy#1262)
* build(deps): bump flate2 from 1.0.21 to 1.0.22 (linkerd/linkerd2-proxy#1263)
* build(deps): bump codecov/codecov-action from 2.0.3 to 2.1.0 (linkerd/linkerd2-proxy#1260)
* trace: remove `/tasks` admin endpoint and `tokio-trace` (linkerd/linkerd2-proxy#1265)
* build(deps): bump softprops/action-gh-release from 0.1.12 to 1 (linkerd/linkerd2-proxy#1264)
* build(deps): bump socket2 from 0.4.1 to 0.4.2 (linkerd/linkerd2-proxy#1268)
* build(deps): bump libc from 0.2.101 to 0.2.102 (linkerd/linkerd2-proxy#1267)
* inbound: don't forward to addresses not in `LINKERD2_PROXY_INBOUND_IPS` (linkerd/linkerd2-proxy#1161)
* build(deps): bump serde_json from 1.0.67 to 1.0.68 (linkerd/linkerd2-proxy#1266)
* inbound: Include TLS labels in authz metrics (linkerd/linkerd2-proxy#1269)
* Update all dependencies (linkerd/linkerd2-proxy#1271)