policy: Do not attempt to validate ServerAuthorizations
The policy-controller's validating admission controller only handles
`Server` resources; but its webhook configuration processes
`ServerAuthorization` resources as well, which results in errors (and
rejections in HA).

This change modifies the webhook configuration so that
`ServerAuthorization` resources are not validated by the admission
controller. This can be restored if the admission controller grows
the ability to validate these resources.

Fixes #6933