Remove new label from mc gateway and service-mirror (#6990)
In #6846 we added the `linkerd.io/extension: multicluster` label to the
gateway and service-mirror pods, for completeness sake, and to use them
in the server podSelector. But it turns out that a deployment's selector
is immutable so we can't have smooth rollouts and this would cause
downtime when upgrading the multicluster resources. So we're removing
those labels and updating the gateway policy accordingly. This also
means we had to add a separate policy for the service-mirror proxy-admin
because we don't have a global label to address the gateway _and_ the
service-mirror through a single saz.