skip prometheus scrapes on policy's `admin`
Fixes #7067

Currently, The policy controller doesn't support prometheus
scrapes on the `admin-http` port under `/metrics` endpoint causing
prometheus to show the target as down.

For now, We can skip the prometheus scrape on the `admin-http` port
by renaming the `admin-http` port to `admin` without having to
update the prometheus config. Once, We start serving the metrics
endpoint, the port can be renamed back.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>