Add tests of multicluster install command
Currently there are no tests of the output of the multicluster install command
unlike the install and viz install commands. This makes it error prone to hard
to validate changes. This is motivated by the addition of an ha mode to the
multicluster components discussed in #7082.

This change adds two test cases and refactors the install command to look like
viz install making it easily testable. This means in practice that the body of
the command is moved into an install function. Here we extract external data,
eg. values, and delegates the values to a render function that handles the
actual rendering.

This is a non-functional change and the output used for the
install_default.golden file is based of the main branch to validate this.

Signed-off-by: Crevil <bjoern.soerensen@gmail.com>