core: use serviceAccount Tokens for pod authentication
Fixes #

This PR updates Linkerd to use a bounded serviceAccount token
through the new volume projection. This token is a separate one, and
not the default (associated with k8s, and externally managed) which
means we can control the expiry, etc on our own and can expect
it to always be present (unlike the defualt, which can be disabled by
the user using `autoMountServiceAccount: false`).

Even though we are using a separate token, we still use the same
service account name as the pod identity (used with metrics, etc) as
these tokens are all generated from the same base service account
attached to the pod (could be defualt, or the user overriden one).
This can be verified by looking at the `user` field in the response.

Example:

Here, The new token was created for the vault audience for a pod
which was using the `mine` serviceAccount in the defualt namespace.

```json
  "kind": "TokenReview",
  "apiVersion": "authentication.k8s.io/v1",
  "metadata": {
    "creationTimestamp": null,
    "managedFields": [
      {
        "manager": "curl",
        "operation": "Update",
        "apiVersion": "authentication.k8s.io/v1",
        "time": "2021-10-19T19:21:40Z",
        "fieldsType": "FieldsV1",
        "fieldsV1": {"f:spec":{"f:audiences":{},"f:token":{}}}
      }
    ]
  },
  "spec": {
    "token": "....",
    "audiences": [
      "vault"
    ]
  },
  "status": {
    "authenticated": true,
    "user": {
      "username": "system:serviceaccount:default:mine",
      "uid": "889a81bd-e31c-4423-b542-98ddca89bfd9",
      "groups": [
        "system:serviceaccounts",
        "system:serviceaccounts:default",
        "system:authenticated"
      ],
      "extra": {
        "authentication.kubernetes.io/pod-name": [
  "nginx"
],
        "authentication.kubernetes.io/pod-uid": [
  "ebf36f80-40ee-48ee-a75b-96dcc21466a6"
]
      }
    },
    "audiences": [
      "vault"
    ]
  }

```

\## Changes

- Update `proxy-injector` and install scripts to include the new
  projected Volume and VolumeMount.
- Update the `identity` pod to validate the token with the linkerd
  audience key.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>