Fix install-pr to use correct registry (#7083)
When PR images are built for integration tests in CI, they use the `ghcr.io` registry. When attempting to install Linkerd from those images using the `bin/install-pr` script, it expects the `cr.l5d.io` registry.

This changes the `install-pr` script to rename those images to use the expected `cr.l5d.io` registry so that users can continue to use a simpler install command—as opposed to passing `--set` flags for _every_ Linkerd image.

`install-pr` is only used for testing and most of the time should be a quick check to make sure some feature is working as expected. This is why I decided renaming made more sense than making the user pass `--set` for each image.

Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>