Add high availability mode to multicluster gateway
Currently the multicluster components cannot be released in a high availability
mode which makes it brittle for the dynamic nature of kubernetes resources.

This change adds an --ha flag to the "linkerd multicluster install" command
which configures the charts to add a PodDisruptionBudget, sets multiple replicas
and anti-affinity on the gateway deployment.

Signed-off-by: Crevil <bjoern.soerensen@gmail.com>