Add control plane namespace arg to policy controller (#7206)
The policy controller synthesizes identity strings based on service account
names; but it assumed that `linkerd` was the name of the control plane
namespace. This change updates the policy controller to take a
`--control-plane-namespace` command-line argument to set this value in
identity strings. The helm templates have been updated to configure the policy
controller appropriately.

Fixes #7204

Co-authored-by: Oliver Gould <ver@buoyant.io>