Remove references to "linkerd-sp-validator" service account
While reviewing the changes for the 2.11 release, I noticed there was still a reference to the "linkerd-sp-validator" service account in the linkerd-psp role binding.

This isn't really an issue as PSPs are deprecated and slated to be removed in Kubernetes 1.25, but figured it would be good to remove the old reference anyway.

Signed-off-by: David Symons <david.symons@onemodel.co>