ci: Split rust test execution into its own step
In the proxy repo, we've split test execution into a distinct step that
follows building the tests (as suggested by [matklad's excellent blog
post][frb]). This allows us to separate the build phase from the text
execution phase (i.e., in the Actions UI).

This change also simplifies the `check` workflow for readability.

[frb]: https://matklad.github.io/2021/09/04/fast-rust-builds.html