Adding the ability to pull identity trust roots from a Secret instead of a ConfigMap (#7345)
Currently, the only way to configure an existing CA is to have an existing ConfigMap or pass in the ca data during installation. For those using cert-manager to generate or retrieve the CA, the ca data will be stored in a Secret.

Adding a boolean installation value (identity.externalCAFromSecret for helm and --identity-external-ca-from-secret for the CLI) allows using an existing Secret as an alternative to an existing ConfigMap

Signed-off-by: Brian Dunnigan <bdun1013dev@gmail.com>